added news feature

+++ backend/stock/models.py
@@ -7,7 +7,7 @@ class StockData(models.Model):
     """
     Model for the stock_stockdata table in Supabase
     """
-    id = models.BigIntegerField(primary_key=True)
+    id = models.AutoField(primary_key=True)
     ticker = models.CharField(max_length=255)
     name = models.CharField(max_length=255)
     current_price = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
